Refactor testerina compiler plugins
